image_stitching_gui/requirements.txt
opencv-python
numpy
Pillow

